Reworked functionality to return highlight snippets per document instead of all matching snippets per search
