+++ solr-facade/src/main/java/com/cherrysoft/solrfacade/mapper/SearchResultMapper.java
@@ -5,8 +5,6 @@ import com.cherrysoft.solrfacade.model.SearchResult;
 import org.mapstruct.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.Collections;
-
 @Mapper(componentModel = "spring")
 public abstract class SearchResultMapper {
   private RecoveredDocumentMapper recoveredDocumentMapper;
@@ -15,7 +13,6 @@ public abstract class SearchResultMapper {
     var recoveredDocumentDtoList = recoveredDocumentMapper.toDtoList(result.getRecoveredDocuments());
     return SearchResultDTO.builder()
         .documents(recoveredDocumentDtoList)
-        .hlSnippets(Collections.unmodifiableList(result.getHighlightSnippets()))
         .spellcheckResult(spellcheckResultDto(result.getSpellcheckResult()))
         .build();
   }
